feat: include election & date in retrieval list
Previously we had the jurisdiction name and round number in the ballot retrieval list name. Now we have the election name (which is user-provided) and the current UTC time to the minute.

This change also improves the "cleaning" done to the name since there are many characters we want to disallow rather than simply spaces. This version allows ASCII letters and numbers, turning everything else into dashes.

Closes #136