refactor(db): exclusively use postgres
Using sqlite for development and testing is easy, but it differs from postgres in ways that make troubleshooting staging and production issues difficult. This commit switches everything from sqlite to postgres in all environments.

The difficulty I had in making this change was primarily around ensuring we're able to easily provide the application with the right configuration in whatever environment it happens to be running in. For example, we need to continue to respect the `DATABASE_URL` environment variable because that's what heroku is using. However, asking people to set environment variables in their shell or IDE to run Arlo properly is not great.

To address this, I set up a system that vaguely resembles Rails with a `config/database.cfg` that is ignored by git but can have per-`FLASK_ENV` database configuration.

Finally, because tests now run against a persisted database we have the possibility of primary key collisions. Since there were lots of primary keys hardcoded in the tests, this happened quite a bit. I fixed this by ensuring that primary keys are always generated randomly, but really we should be using something like https://github.com/jeancochrane/pytest-flask-sqlalchemy to automatically roll back sessions for each test. This was more difficult to set up than I expected, so I'm leaving it as a future task.

Closes #199