feat: add client and server vote total validation
This could still be better, i.e. all client validation should be repeated on the server, but this is a start. Now "Vote for 3" style elections can be better validating by checking the the votes for a contest do not exceed the product of ballots times contest votes per ballot.